requirements.txt
cryptography==41.0.3
Flask==3.0.3
argon2-cffi>=25.1.0
PyQt6
requests
coverage
PyInstaller
